Move async data fetching to themes/controller
